+++ src/expression/rle_expression.cpp
@@ -214,7 +214,8 @@ void dec_rle_map_opr<FlsString, INDEX_PT>::Decode(n_t              vec_idx,
 
 		if (byte_arr_vec.capacity() - byte_arr_vec.size() < CFG::String::max_bytes_per_string) {
 			byte_arr_vec.reserve(
-			    std::max(byte_arr_vec.capacity() * 2, byte_arr_vec.size() + CFG::String::max_bytes_per_string));
+			    std::max(byte_arr_vec.capacity() * 2,
+			             byte_arr_vec.size() + static_cast<size_t>(CFG::String::max_bytes_per_string)));
 		}
 		byte_arr_vec.insert(byte_arr_vec.end(), bytes + cur_ofs, bytes + next_offset);
 	}
